Move code to mc trial
