+++ hoomd-vector/src/quaternion.rs
@@ -3,7 +3,6 @@
 
 //! Implement [`Quaternion`] and related types.
 use std::{
-    f64::consts::PI,
     fmt,
     ops::{Add, AddAssign, Div, DivAssign, Mul, MulAssign, Sub, SubAssign},
 };
@@ -13,7 +12,7 @@ use rand::{
     Rng,
     distr::{Distribution, StandardUniform},
 };
-use rand_distr::{Normal, StandardNormal};
+use rand_distr::StandardNormal;
 
 use crate::{Cartesian, Cross, Error, InnerProduct, Rotate, Rotation, RotationMatrix, Unit};
 
@@ -784,88 +783,6 @@ impl Distribution<Versor> for StandardUniform {
             scalar: scalar / norm,
             vector: vector / norm,
         })
-    }
-}
-
-/// A normal distribution of random Versors, centered on a mean with
-/// some standard deviation.
-#[derive(Debug, Clone, Copy)]
-pub struct VersorDisplacement {
-    /// The center of the distribution on the 3-Sphere.
-    mean: Versor,
-    /// The standard deviation of the normal distribution of quaternions around the mean.
-    // normal: Normal<f64>,
-    std_dev: f64,
-}
-
-impl From<(Versor, f64)> for VersorDisplacement {
-    #[inline]
-    fn from(value: (Versor, f64)) -> Self {
-        Self {
-            mean: value.0,
-            std_dev: value.1,
-        }
-    }
-}
-impl Distribution<Versor> for VersorDisplacement {
-    /// Sample a random [`Versor`] displacement from a provided mean.
-    ///
-    /// Mathematically, we sample from a 3-dimensional Normal distribution
-    /// in the tangent space of SO(3), lift to the manifold, then rotate to center on
-    /// the mean of the [`VersorDisplacement`]. The result is a small displacement from
-    /// a quaternion input, with fast decay in the tails that make large displacements
-    /// unlikely. This is desirable for Monte Carlo, as large moves are very likely to
-    /// be rejected.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// use hoomd_vector::{Versor, VersorDisplacement};
-    /// use rand::{Rng, SeedableRng, rngs::StdRng};
-    ///
-    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
-    /// let mut rng = StdRng::seed_from_u64(1);
-    /// let normal = VersorDisplacement::from(Versor::default(), 0.1);
-    /// let v: Versor = normal.sample(&mut rng);
-    /// # Ok(())
-    /// # }
-    /// ```
-    #[inline]
-    fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> Versor {
-        // Based on Karney 2007: doi.org/10.1016/j.jmgm.2006.04.002
-        const _SAFE_THETA: f64 = 1e-12;
-        loop {
-            // As in section 7, we select `s` from a 3D Gaussian distribution.
-            let normal =
-                Normal::new(0.0, self.std_dev).expect("Failed to create normal distribution.");
-
-            let s = Cartesian::from([(); 3].map(|()| normal.sample(rng)));
-            let theta = s.norm();
-
-            // Reject moves with |s| > pi to ensure detailed balance. Otherwise, there
-            // is a shorter path between the proposed move and the start (theta - pi),
-            // which has a different rejection probability.
-            if theta > PI {
-                continue;
-            }
-
-            // Lift the normally distributed values to SO(3) with the exponential map
-            let half_theta = 0.5 * theta;
-            let w = half_theta.cos();
-
-            // If theta is near 0, do not compute the value. TODO: detailed balance?
-            let v_factor = if theta < _SAFE_THETA {
-                0.5
-            } else {
-                half_theta.sin() / theta
-            };
-
-            let v = s * v_factor * half_theta.sin();
-
-            // We are normalized by construction, so call the tuple initializer
-            // Then, rotate by the current quaternion we displace from
-            return Versor(Quaternion::from([w, v[0], v[1], v[2]])).combine(&self.mean);
-        }
     }
 }
